requirements.txt
flask==1.0.2
flask-sqlalchemy==2.3.2
python-decouple==3.1
Flask-Migrate==2.3.0
Flask-RQ2==18.1
redis==2.10.6
requests==2.20.1
flask-marshmallow==0.9.0
